backend/requirements/development.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
bleach==6.2.0
celery==5.6.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.27
django-admin-sortable2==2.3.1
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-js-asset==2.2.0
django-ratelimit==4.1.0
django-resized==1.0.3
django-smart-selects==1.7.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
kombu==5.6.2
packaging==26.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
redis==7.0.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.5.3
webencodings==0.5.1
